requirements.txt
Pillow==10.4.0
pyirsdk==1.3.5
pypresence==4.3.0
pystray==0.19.5
